python/pyproject.toml
[tool.poetry]
name = "su-expa-final-references-python"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "app/"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
click = "^8.1.6"
aiofiles = "^23.1.0"
aiocsv = "^1.2.4"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.278"
ipython = "^8.14.0"
pyright = "^1.1.317"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
skip-string-normalization = false  # ''を""へ直すのをスキップするか
include = '\.py$'
exclude = '''
(
  \.eggs
| \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| \.idea
| \.vscode
| _build
| buck-out
| build
| dist
)
'''

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120
